requirements.txt
torch
torchvision
tqdm
Pillow
numpy
matplotlib
yaml==5.3.1
tensorboardx==2.1

